Update Kafka version to 6.0.0.
This commit includes changes that are necessary to fix the issues with the
latest Kafka 6.0.0 / 2.6.0 versions.

- Added implicit serializer for the test Avro record.

- Fixed issue with Schema Registry setup for each test

Using beforeEach and afterEach for setting up the Schema Registry fails with the
following error:

```
SerializationException: Error retrieving Avro unknown schema for id 1
...
Schema id 1 not found with cached schema registry rest client
```

Now we start the EmbeddedKafka stack for each integration test suite and use
different topic name for each test so that the tests are isolated.

- Updated main Scala version to 2.12.11

This is required since the newest Kafka releases are compiled against the
`2.12.11` version. Using the latest Scala 2.12.12 version fails with
compatibility issue.