Allow the specification of the offset reset strategy (#26)
When no previous offset is present (because it is the first load) or the last offset gathered from the table is out of range (because the last load was longer ago than the topic retention time), the connector always starts from the end of topic (kafka default strategy 'latest'). For low throughput topics there might be no message at that point of time so again, no new offset is stored and the messages are lost. Additionally, for topics that work with compaction it is expected to load everything from beginning the topic.

This commit allows specifying `AUTO_OFFSET_RESET` as property and the value is passed to the kafka consumer.

Co-authored-by: Schulte, Johannes <johannes.schulte@eventim.de>
Co-authored-by: Muhammet Orazov <m.orazow@gmail.com>